refactor : request를 도메인으로 매핑하는 역할을 controller가 아닌 service가 맡도록 수정
