+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -30,8 +30,7 @@ public class ReservationTimeController {
 
     @PostMapping("/times")
     public ResponseEntity<ReservationTime> addReservationTime(@RequestBody ReservationTimeRequest request) {
-        ReservationTime reservationTime = reservationTimeService.addReservationTime(
-                new ReservationTime(request.startAt()));
+        ReservationTime reservationTime = reservationTimeService.addReservationTime(request);
         return ResponseEntity.created(URI.create("/times/" + reservationTime.getId())).body(reservationTime);
     }
